requirements.txt
absl-py==0.13.0
alabaster==0.7.12
alembic==1.6.5
appdirs==1.4.4
appnope==0.1.2
APScheduler==3.7.0
argon2-cffi==20.1.0
arrow==1.1.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
Babel==2.9.1
backcall==0.2.0
binaryornot==0.4.4
black==21.4b0
bleach==3.3.0
Brotli==1.0.9
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
chex==0.0.8
click==7.1.2
cliff==3.8.0
cloudinary==1.26.0
cmaes==0.8.2
cmd2==2.1.2
codecarbon==1.2.0
colorama==0.4.4
colorlog==6.4.1
commonmark==0.9.1
cookiecutter==1.7.2
cycler==0.10.0
dash==1.21.0
dash-bootstrap-components==0.13.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
datasets==1.11.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.4
dm-tree==0.1.6
docopt==0.6.2
docutils==0.16
entrypoints==0.3
execnet==1.9.0
faiss-cpu==1.7.1.post2
filelock==3.0.12
fire==0.4.0
flake8==3.9.2
flashtext==2.7
Flask==1.1.2
Flask-Compress==1.10.1
Flask-Cors==3.0.10
flatbuffers==1.12
flax==0.3.4
fsspec==2021.7.0
fugashi==1.1.1
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.32.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
hdf5storage==0.1.18
huggingface-hub==0.0.12
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
ipadic==1.0.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.9.3
itsdangerous==1.1.0
jax==0.2.19
jaxlib==0.1.70
jedi==0.18.0
Jinja2==2.11.3
jinja2-time==0.2.0
joblib==1.0.1
jsonable==0.3.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.3.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keras2onnx==1.7.0
kiwisolver==1.3.1
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.4.2
mccabe==0.6.1
mediawiki-utilities==0.4.18
mistune==0.8.4
msgpack==1.0.2
multiprocess==0.70.12.2
mwapi==0.5.1
mwtypes==0.3.2
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.2.0
numpy==1.19.5
oauthlib==3.1.1
onnx==1.10.1
onnxconverter-common==1.8.1
opencv-python==4.5.1.48
opt-einsum==3.3.0
optax==0.0.9
optuna==2.9.1
packaging==20.9
pandas==1.1.3
pandocfilters==1.4.3
parameterized==0.8.1
parso==0.8.1
pathspec==0.9.0
pbr==5.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plac==1.3.3
plotly==5.2.2
pluggy==0.13.1
portalocker==2.0.0
poyo==0.5.0
prettytable==2.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.17
protobuf==3.17.3
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
py-cpuinfo==8.0.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pygame==2.0.2
Pygments==2.8.1
PyMySQL==1.0.2
pynvml==11.0.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.17.3
pytest==6.2.4
pytest-forked==1.3.0
pytest-timeout==1.4.2
pytest-xdist==2.3.0
python-dateutil==2.8.1
python-dotenv==0.19.1
python-editor==1.0.4
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
ray==1.6.0
recommonmark==0.7.1
redis==3.5.3
regex==2021.8.3
requests==2.25.1
requests-oauthlib==1.3.0
rouge-score==0.0.4
rsa==4.7.2
sacrebleu==1.5.1
sacremoses==0.0.45
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
sentencepiece==0.1.94
six==1.15.0
sklearn==0.0
smmap==4.0.0
snowballstemmer==2.1.0
SoundFile==0.10.3.post1
Sphinx==3.2.1
sphinx-copybutton==0.4.0
sphinx-intl==2.0.1
sphinx-markdown-tables==0.0.15
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxext-opengraph==0.4.1
SQLAlchemy==1.4.23
stevedore==3.4.0
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.9.3
testpath==0.4.4
text-unidecode==1.3
threadpoolctl==2.1.0
timeout-decorator==0.5.0
timm==0.4.12
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.60.0
traitlets==5.0.5
-e git+https://github.com/Apoorvgarg-creator/transformers.git@0f798fb6cc072f55bdc47b80afb2a51740e3acac#egg=transformers
typing-extensions==3.7.4.3
tzlocal==2.1
unidic==1.0.3
unidic-lite==1.0.8
urllib3==1.26.4
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xxhash==2.0.2
yamlconf==0.2.4